Initial commit of the Phoenix Proof-of-concept openSIL codebase
This is the initial release of openSIL proof-of-concept codebase for the
Phoenix Zen 4 APU (Family 19h Model 70h). This release targets both
mobile (Package FP7/FP7r2) and desktop (Socket AM5) chips.

Note that this codebase only supports Phoenix and Hawk Point processors.
It does not Phoenix2.

See the Public PPR here:
https://docs.amd.com/v/u/en-US/57019-A0-PUB_3.00
or the archive here:
https://web.archive.org/web/20240807200252/https://www.amd.com/content/dam/amd/en/documents/processor-tech-docs/programmer-references/57019-A0-PUB_3_00.zip

Signed-off-by: Martin Roth <martin.roth@amd.com>